pyproject.toml
[tool]
[tool.poetry]
name = "ampoule-cli"
version = "0.1.0"
description = ""
authors = ["Lucas Denefle <lucas.denefle@converge.io>"]
packages = [{include = "ampoule_cli"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
pyserial = "^3.5"
protobuf = "^4.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
